backend/requirements.txt
Django==4.2
django-filter==25.1
djangorestframework==3.16.0
djoser==2.3.1
Pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1